refactor: modularize button component and improve structure
Moved Button styles and interfaces into separate files for better modularity and maintainability. Refactored the component to rely on a centralized `buttonStyle` object for styling logic, reducing duplication. Adjusted stories to align with the updated structure and to include additional test coverage for button variations.
